requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
cffi==1.16.0
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
jwt==1.3.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.4.4
uuid==1.30
django-filter==23.2